requirements.txt
akshare>=1.13.0
pandas>=1.5
numpy>=1.23
PyYAML>=6.0
python-dateutil>=2.8.2
